requirements.txt
# AI/機械学習ライブラリ
tensorflow==2.15.0
numpy>=1.24.0
pillow>=10.0.0

# 画像処理
imagehash>=4.3.1

# ユーティリティ
requests>=2.31.0
beautifulsoup4>=4.12.0

# オプション: より高度な画像検索
# bing-image-downloader>=1.1.2
# google-images-download>=2.8.0